ANDROID: iwl7000: pass full FW info to transport
The code currently passes only the specific image that should
be loaded, but then has to pass the IML (image loader) out of
band, which is confusing. Pass the full FW data together with
desired image type, and use the IML from that.

This also cleans up the code in the various sub-drivers a bit
as they no longer have to look up and check for the image.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: Iebadc56bb2762e5f4d71f66bb2609d74b33daf11
iwl7000-tree: 356c5368d60e9c9ed34a872b8768f4f60149fd4c