ANDROID: iwl7000: mac80211: fix deactivated link CSA
If the link is deactivated and the CSA completes, then that
needs to update the link station's bandwidth (only the AP STA
can exist at this point, no TDLS on inactive links) and set
the CSA to no longer be active. Fix this.

Change-Id: I5a868c501ee73fcc2355d61c2ee06e5f444b350f
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 3fc46881882b8c1b1964ec27bd1df687394110af