ANDROID: iwl7000: rename struct iwl_base_params
These are (going to be) base MAC parameters that are identical
even for different platforms with the same MAC, so rename the
structure accordingly, calling it iwl_family_base_params.

Also rename the pointer to it so the dereferencing is a bit
shorter.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: I1108eb51a4935abec4206f0aac461a3833033e79
iwl7000-tree: 821040c8ff39010c828274a64c1bf1daa9b27a8e