ANDROID: iwl7000: cfg: remove MAC type/step matching
Now that it's all split into MAC and RF configs, remove
the matching on MAC type and step. If we ever need to do
something based on the MAC step, we'll have to find some
new mechanism (since the MAC type is known already from
the PCI IDs table, but not the step), or just handle the
(likely small) differences in code.

Change-Id: Iae27b781221fd29845493adf2c29d9e4f7a9c33b
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 8ac1468729f906b06e25ea106284cb8d5b3187bc