ANDROID: iwl7000: pcie: avoid -Woverflow warning/error
When none of the sub-drivers are built, the ARRAY_SIZE()
will be zero and be caught by the previous runtime check,
but the compiler might still emit an overflow warning.
Add an explicit cast to silence it, if it's really zero
the previous check means the loop won't execute anyway.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: I9df3bd841cad346a37258ce4d0a649691bec902d
iwl7000-tree: 972f1cdc9643ff46a5e843e9bd15f645bedc9f33