ANDROID: iwl7000: mld: avoid init-after-queue
rx_omi::finished_work is initialized when the containing link is.
If the worker was queued and then an error happened, we will get to
iwl_mld_init_link from the reconfig and initialize the work after it was
queued.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ie12a2968563ea088b4884f139acea59060a51871
iwl7000-tree: 9641a072b9dc417ab663c9f6436a1f17e2ac86e6