ANDROID: iwl7000: pcie: move generation specific files to a folder
As a new generation of pcie is going to be written, we will need a
folder for each generation.
Since gen1 and gen2 code is tightly coupled and has with shared logic -
it is not really separable.
Put the code of both in one folder.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I66345c2b3fda55dcb8ff779c64de72d5c19f6649
iwl7000-tree: a6fafaedb5d60956f8dc4d0cf248fff9e879e733