ANDROID: iwl7000: mld: allow 2 ROCs on the same vif
In the current code, if an ROC is started on a vif that already has an
active ROC we reject it and warn.

But really there is no such limitation. The actual limitation is to not
have 2 ROCs of the same type simultaneously.

Add a helper function to find a vif that has an active ROC of a given
type, and only if one exist - reject the ROC.

This allows also to remove bss_roc_vif.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I17cb191596ed4e97a4854108f8ca5ca197662a62
iwl7000-tree: d845616c3c31fb9096d0e5fe08f9f56d2e48977a