ANDROID: iwl7000: pcie: add missing TOP reset code
The TOP reset requires code to handle the interrupt, which had
been in my patch at some point, but clearly got lost. As the
test had been running on the wrong hardware and failing due to
that, we missed this. Add the missing code.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: If138ceeb8bb178b3931d96b537f746346227e681
iwl7000-tree: 20f395902a0a9071139f4517197721e1c9a69372