ANDROID: iwl7000: mld: don't leak FW link IDs
On restart, we will re-add all the links we had before it,
which also means that we re-allocate a fw id for those links.
But we never free the fw id that was allocated to the link from before
the restart.

On restart, no link exists in the fw, so also the fw ids should be
released. Do this when cleaning up the link.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I2a7cef4698b935597032f2798d8d8914e8e65bf3
iwl7000-tree: 1d162d45426cc49bea01ba1a31033f6add4452a2