ANDROID: iwl7000: ensure iwl_average_neg_dbm is available for kunit
The function was not compiled if CONFIG_INET was not enabled. Change it
to always be compiled in. Also update the IWLWIFI_OPMODE_MODULAR check
so that the symbol is exported even if only the kunit tests are build
as a module.

Change-Id: I02014d338de26768533cd07225780563393b4591
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202505120905.ceWoI4rO-lkp@intel.com/
Fixes: 1f52f7f0abe2 ("wifi: iwlwifi: move dBm averaging function into utils")
iwl7000-tree: 800fe085168e073910d8a0b758df0399d3c38323