ANDROID: iwl7000: gen3: implement NIC ownership acquisition for gen3
In case of vPro platforms, the ownership can be shared and taken between
the host and ME. This is done by the host sending a request to the
ME, which then responds with the ownership status.

The patch implements the ownership acquisition
and the sw reset flow which is done afterwards, for gen3 devices.

Signed-off-by: Itamar Shalevit <itamar.shalev@intel.com>
Change-Id: I0d3c4f38905555e82d408e7cb0776e6b1e419172
iwl7000-tree: 6714bd1b75ada4b3a90956001b2962f7d94d11bd