ANDROID: iwl7000: match discrete/integrated to fix some names
Some device names were wrong because our internal data suggested
that discrete Ga devices have B-step RF, when they actually have
C-step. However, matching the step for them is bad anyway.

Change the code to be able to find the devinfo depending on the
device being integrated or discrete. This is only for the names,
since the RF config cannot be different for the same RF because
it's discrete or integrated, so add a kunit test that ensures
both (a) the RF config is the same and (b) the name is different
(the latter really only because that's the whole point of having
a match on the discrete/integrated bit.)

Remove the RF step matching since it's no longer needed now.

Change-Id: Ie5919c70e9d8e3a28152aaf3cdffd19ed3d4f5c7
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 3ac861a9f61158fd11ea5e87198beac37dc9cd56