ANDROID: iwl7000: fix HE/EHT capabilities
The default capabilities were set to much more than the hardware
currently is intended to support, and then masked off for only
the GL MAC type. However, this was due to some miscommunication
and is incorrect, it should've been masked off for all current
and planned MACs/RFs. Instead of doing this removing and adding
of capabilities, simply list the ones that currently apply to
all devices. If this changes for a new device type we can change
the code, but at least the driver won't erroneously advertise
some capabilities that aren't actually implemented in hardware.

Change-Id: I7885499bc80453d5a84285ec927f5e89f9adfb21
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: c9656ad5113db939807710b4896a3342fb6ed596