ANDROID: iwl7000: fix SO/JF FW API range
SO was frozen on API 89, and JF was frozen on API 77.
My previous patch bumped the minimum API of SO to 89.
While this makes sense, it breaks JF: we require 89 FW but such doesn't
exist, of course.

To fix it, define the API range also for JF (to [77, 77]),
and change the logic of selecting the API range to handle also
non-overlapping ranges - by simply taking the lower range of both
mac range and crf range.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I169527762359634b7e4887c080df206e1a3dcc30
iwl7000-tree: b23d6d3c93f22081a00bf1ba29c42b9ce331a28d