ANDROID: iwl7000: mld: use the right macro for KUNIT visibility
Use the dedicated VISIBLE_IF_IWLWIFI_KUNIT and not the general
VISIBLE_IF_KUNIT to make a function static or not, depending on the
config.
This fixes an issue in which IWLWIFI_KUNIT_TESTS is not set, but KUNIT
is. In that case, iwl_mld_emlsr_pair_state will be non-static but won't
have a prototype.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ia6fb3dc69d4f434e885b9ec05f74fc2657f9e2be
iwl7000-tree: e57789ffa99633a151e7bb170f6f20f7f216d1ab