ANDROID: iwl7000: mac80211: only assign chanctx in reconfig
At the end of reconfig we are activating all the links that were active
before the error.
During the activation, _ieee80211_link_use_channel will unassign and
re-assign the chanctx from/to the link.
But we only need to do the assign, as we are re-building the state as it
was before the reconfig.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ia5f68992c7c112bea8a426c9339f50c88be3a9ca
iwl7000-tree: f303502572ff1875f7142db71b5782d9aee2e160