ANDROID: iwl7000: mac80211: don't require cipher and keylen in gtk rekey
ieee80211_add_gtk_rekey receives a keyconf as an argument, and the
cipher and keylen are taken from there to the new allocated key.
But in rekey, both the cipher and the keylen should be the same as of
the old key, so let ieee80211_add_gtk_rekey find those, so drivers won't
have to fill it in.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ie6594ae2b4b6d5b3d536e642b349046ebfce7a5d
iwl7000-tree: 221691d7624a2ce73abd6125f0f2a2f030cb64a3