ANDROID: iwl7000: mld: cleanup cipher lookup in resume
We used to lookup the ciphers of the mcast keys, but this was beacuse it
was required for ieee80211_get_rekey_add. Now as this API no longer
needs the cipher as an argument, we can remove the cipher lookups.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Iae55243b575248cb4cc0b416f7f63092b5803219
iwl7000-tree: bbf07fc55be7d3b325deed6130654372e249c5c4