ANDROID: iwl7000: mvm: cleanup cipher lookup in resume
We used to lookup the ciphers of the mcast keys, but this was beacuse it
was required for ieee80211_get_rekey_add. Now as this API no longer
needs the cipher as an argument, we can remove the cipher lookups.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I5d92258a9d63a39ee3acb02a72a2af9984993018
iwl7000-tree: a157a1bb4ab5aa6053a9227f491a2517740235dd