ANDROID: iwl7000: don't read SMEM on CC as well
This is a large amount of data, so it slows down the reset, which causes
ping loss.
We implemented this for HR but the same issue started to pop up on GF as
well, so apply this workaround for GF as well.

Change-Id: I9298d24f20b7c0490ebc19b9914a0f590907983c
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: ef1b4617613414da5d996360c7b5049f814ae41c