ANDROID: iwl7000: make ppag versioning clear
We used to have in iwl_ppag_table_cmd v2, that covered multiple FW
versions of the command (2-6), so we just called it v2, and v3 for FW
version 7. This is a bit confusing, and now v2 actually covers only FW
version 5. Rename v2 to v5 and v3 to v7 so we don't have a different
versioning than the FW has.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: If7898bdf9ef70eed9c12484c03a3cc4f27635682
iwl7000-tree: c90851fbb42a7c0ca9900d179d30409bf98cef52