ANDROID: iwl7000: mld: copy data for untrusted devices
The IOMMU synchronous flush for untrusted devices is _really_
expensive (75-80% CPU usage), in part because the driver has
a high degree of parallelism, but the IOMMU goes through a
single spinlock. Given that, it's actually cheaper to copy
the whole data into a fresh SKB allocation and keep the page
mapped for the device.

Doing this requires being careful and not accessing the data
multiple times while checking since it can still be modified
by the device (at least in theory), so copy the len_n_flags,
packet header and MPDU descriptor to the stack.

Change-Id: I030456b4585cab5c5d9ec0ae955c427212caf236
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 6d926b950aeb70de7628f75305a29af51c11305b