ANDROID: iwl7000: fix new API scheme for chrome
In Chrome we don't support a FW that is older than the driver.
This means that if the first attept to load a FW failed, the init will
fail.

In addition, we are changing the API scheme from API = CORE + 3
to API = c<CORE>.
We are currently in a stage where the driver supports the new
scheme, but there are still no such FW files.

So the first attempt will fail. For non-chrome,
we will fallback to the old API scheme of the *same core* (99)
and it will be fine.

But for Chrome, as said, we don't allow any fallback.

Change the code to allow only the fallback from core99-new-scheme
to core99-old-scheme.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Idedfdd2bd652acff45b59b0b9dee4a3ffa953de1
iwl7000-tree: 48cd413ba5372546d4b74eeacac529bbd2daa10c