ANDROID: iwl7000: carefully select the PNVM source
For newer device, and from API 100 (core 97), the PNVM should be taken
from the .ucode file, and not from an external .pnvm file.

In the current logic, if the PNVM doesn't exist in the .ucode file, we
fallback to fetching the .ucode file. This is wrong and hides bugs.

This fallback was needed for (a) old devices and (b) for newer
devices with an old API.

Since we no longer support those old APIs, (b) is not longer relevant.
We can, according to the device, select the right PNVM source
and fail if we couldn't find the PNVM there.

Add clear logic to select the expected PNVM source, and print an error
if we couldn't get the PNVM from there.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I87fbcd25bbee733d2612206b76c2d8593d0cbd39
iwl7000-tree: e1123c5c3bd1d50ee1ea6a4a91c75ef604bbb2d1