ANDROID: iwl7000: xvt: remove WARN_ON for messages to userspace
If (for example) the userspace netlink socket size isn't sufficient
for the messages XVT wants to send, we hit this WARN_ON(). But it's
not useful for the driver itself, more of a test canary, but tests
are now allow-listing this WARN_ON() instead. Remove it instead of
doing that in more tests.

Change-Id: I5b0d8611951236d1c1f68fdd2b4a7d0c95189e69
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 4c69cf28f9632dee720c6038b4beb707c8ad5b3b