ANDROID: iwl7000: pcie: remember when interrupts are disabled
trans_pcie::fh_mask and hw_mask indicates what are the interrupts are
currently enabled (unmasked).
When we disable all interrupts, those should be set to 0, so if, for
some reason, we get an interrupt even though it was disabled, we will
know to ignore.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I919375e5ad7bd7e4fee4a95ce6ce6978653d6b16
iwl7000-tree: d49e9edac49f3bd36598ea9027279789b520f426