ANDROID: iwl7000: mld: support TLC command version 5
A new version of the TLC command was added in order to support the new
MCSs intoduced in UHR, and an indication of ELR support.

To support the new MCSs, the new version will have MCS bitmaps
(ht_rates) of 32 bit and not 16 bit, as in the old version.

Change the code to populate the new version of the command,
and if the FW requires the old version, copy the content of the new version
structure to the old version structure.

Note that this doesn't actually set the new MCSs, this will come later.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Iecf6570c9fe11d8fbdc0718341ac92506b02d78c
iwl7000-tree: d5539a1a358d4baaedb22ad4d152184be2c98fd0