ANDROID: iwl7000: pcie: move pm_support to the specific transport
Currently it is under iwl_trans, which is the bus agnostic part of the
transport. But really it is relevant for pcie only, so move it to the
iwl_trans_pcie and export it via an API to the opmode.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ic6648f040430c94150d0fa11601f50a6a630b862
iwl7000-tree: cfdfdbc8994a2f7dba1ee3e641175981b421d1a7