ANDROID: iwl7000: pcie: move ltr_enabled to the specific transport
Currently it is under iwl_trans, which is the bus agnostic part of the
transport. But really it is relevant for pcie only, so move it to the
iwl_trans_pcie and export it via an API to the opmode.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I3eeca8b17abeba6ed30f0d681518c81ede0acf30
iwl7000-tree: c6384f00844804b2ac7d541ac9b4a894596d332e