ANDROID: iwl7000: mld: support testmode
In case we receive a command which the driver doesn't support,
we check if the command is handled in the cmd_exec callback.
If not, (and after a few more checks) it'll return -EOPNOTSUPP,
as required.

But in order to do those checks, we need the testmode object to be
initialized, or we get a NULL pointer dereference.

Call iwl_tm_init also in iwlmld to avoid that.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I9d3071fabc7e12cb2a2b60480dbb9419a630fcf7
iwl7000-tree: 3f5cfd706fd085894dfd2810e6b1977d5ca157bd