ANDROID: iwl7000: radiotap: add a field for PHY SERVICE field related data
For certain frames bits are encoded into the DATA field of the PHY
header in a way that cannot be properly reported in other ways. Add a
new field to be able to report the SERVICE field and scrambler
initialization value as well as the bandwidth signalling information
that may be encoded here.

Change-Id: Ic11e92b56b2b1df6c202ec9442f25bd78206a4e8
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: a718ff65584687e4a0c0c647ca54ddd1b158ebba